youtube-search/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
pytube==15.0.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-pinecone==0.0.1
pinecone-client==3.0.1
python-multipart==0.0.6
httpx==0.26.0
pydantic==2.5.3
openai==1.7.1
tiktoken==0.5.2
tenacity==8.2.3
typing-extensions==4.9.0 